requirements.txt
tensorflow
opencv-python
Pillow
matplotlib
pandas
numpy
pyautogui
